webmessaging/postMessage_cross_domain_image_transfer.sub.htm

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web Messaging - cross-domain image transfer</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script type="text/javascript" charset="utf-8">
function PostMessageTest() {
  var t = async_test("Cross-domain image transfer");

  // Create a gray image.
  var size = 4
  var sourceCanvas = document.createElement('canvas');
  sourceCanvas.width = sourceCanvas.height = size;
  var gl = sourceCanvas.getContext('webgl');
  gl.viewport(0, 0, gl.drawingBufferWidth, gl.drawingBufferHeight);
  gl.clearColor(0.5, 0.5, 0.5, 1.0);
  gl.clear(gl.COLOR_BUFFER_BIT);

  var expected_pixels = "[128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255,128,128,128,255]"

  addEventListener('message', t.step_func(function(e) {
      assert_equals(e.data, expected_pixels)
      t.done()
  }));

  createImageBitmap(sourceCanvas).then( bitmap => {
    document.getElementById("receiver").contentWindow.postMessage({image: bitmap}, "*", [bitmap]);
  });
}
</script>


<body>
    <h1>Web Messaging - cross-domain image transfer</h1>
    <iframe id="receiver" onload="PostMessageTest()" src="{{location[scheme]}}://{{domains[www]}}:{{location[port]}}/webmessaging/support/cross-domain-image-receive.htm"></iframe>
</body>
</html>
